chore(proposals): refactor components structure
