+++ src/app/proposals/components/HeaderSection.tsx
