+++ src/app/proposals/components/LatestProposalsTable.tsx
@@ -7,16 +7,16 @@ import {
   useReactTable,
   getPaginationRowModel,
 } from '@tanstack/react-table'
-import { type LatestProposalResponse } from './hooks/useFetchLatestProposals'
-import { StatusColumn } from '@/app/proposals/table-columns/StatusColumn'
+import { type LatestProposalResponse } from '../hooks/useFetchLatestProposals'
+import { StatusColumn } from './table-columns/StatusColumn'
 import { StatefulTable } from '@/components/Table'
 import { HeaderTitle, Typography } from '@/components/Typography'
-import { SentimentColumn } from '@/app/proposals/table-columns/SentimentColumn'
-import { ProposalNameColumn } from '@/app/proposals/table-columns/ProposalNameColumn'
+import { SentimentColumn } from './table-columns/SentimentColumn'
+import { ProposalNameColumn } from './table-columns/ProposalNameColumn'
 import { VotesColumn } from './table-columns/VotesColumn'
 import { TimeColumn } from './table-columns/TimeColumn'
-import { DebounceSearch } from '../../components/DebounceSearch/DebounceSearch'
-import { useProposalListData } from './hooks/useProposalListData'
+import { DebounceSearch } from '@/components/DebounceSearch'
+import { useProposalListData } from '../hooks/useProposalListData'
 import { Button } from '@/components/Button'
 
 interface LatestProposalsTableProps {
