+++ src/app/proposals/components/table-columns/SentimentColumn.tsx
