+++ src/app/proposals/components/table-columns/VotesColumn.tsx
