+++ src/app/proposals/page.tsx
@@ -1,7 +1,7 @@
 'use client'
-import { HeaderSection } from '@/app/proposals/HeaderSection'
+import { HeaderSection } from './components/HeaderSection'
 import { useFetchAllProposals } from '@/app/proposals/hooks/useFetchLatestProposals'
-import { LatestProposalsTableMemoized } from '@/app/proposals/LatestProposalsTable'
+import { LatestProposalsTableMemoized } from './components/LatestProposalsTable'
 import { MetricsCard } from '@/components/MetricsCard'
 import { Popover } from '@/components/Popover'
 import { TxStatusMessage } from '@/components/TxStatusMessage/TxStatusMessage'
